+++ modules/types/src/index.ts
@@ -18,3 +18,5 @@ export type {
   NumberArray16
 } from './array-types';
 export {isTypedArray, isNumericArray} from './is-array';
+
+export type {Bounds, Bounds2D, Bounds3D} from './bounds-types';