+++ docs/whats-new.md
@@ -45,11 +45,22 @@
   </tbody>
 </table>
 
+## v4.1
+
+Release Date: Q3, 2024.
+
+- [SPDX](https://spdx.org/licenses/) license headers.
+
+**`@math.gl/core`**
+
+- New `Vector2Like` - `Matrix4Like` types to specify numeric inputs of a specific length.
 
 **`@math.gl/types`**
 
 - New types for expressing [bounds](./modules/types/api-reference/bounds) (extents): `Bounds`, `Bounds2D` and `Bounds3D`.
-- 
+- New types to specify numeric arrays of a specific length: `NumberArray2` - `NumberArray16`. 
+- `isTypedArray()` and `isNumericArray()` utilities now perform typescript type narrowing.
+
 ## v4.0
 
 Release Date: Oct 14, 2023.